fix rendering for reports that are in-progress/failed
